requirements.txt
azure-core==1.26.1
azure-functions==1.12.0
azure-storage-blob==12.14.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
flake8==5.0.4
flatten-dict==0.4.2
idna==3.4
isodate==0.6.1
marshmallow==3.19.0
mccabe==0.7.0
msrest==0.7.1
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.12
xmltodict==0.13.0
